Update LC branch

+++ MovieTopia/MaintainTheatres.cs
@@ -15,6 +15,7 @@ namespace MovieTopia
     {
         private string DATABASE_URL;
         private int padding = 20;
+        private string tblName = "Theatre";
         DataSet ds;
         SqlDataAdapter adapter;
 
@@ -31,13 +32,19 @@ namespace MovieTopia
 
         private void Form_Resize(Object sender, EventArgs e)
         {
+            lblName.Top = padding / 2;
             lblName.Left = (this.ClientSize.Width - lblName.Width) / 2;
             btnEdit.Left = (this.ClientSize.Width - btnEdit.Width) / 2;
             btnNew.Left = btnEdit.Left - btnEdit.Width - padding;
             btnDelete.Left = btnEdit.Left + btnEdit.Width + padding;
+            btnReturn.Left = this.ClientSize.Width - btnReturn.Width - padding;
             btnNew.Top = (this.ClientSize.Height - btnEdit.Height - padding * 2);
             btnEdit.Top = (this.ClientSize.Height - btnEdit.Height - padding * 2);
             btnDelete.Top = (this.ClientSize.Height - btnDelete.Height - padding * 2);
+            btnReturn.Top = (this.ClientSize.Height - btnDelete.Height - padding * 2);
+
+            lblFilter.Location = new Point(padding, 3 * padding);
+            txtFilter.Location = new Point(lblFilter.Left + lblFilter.Width + padding, 3 * padding);
 
             AdjustDataGridViewSize();
             AdjustColumnWidths();
@@ -56,19 +63,20 @@ namespace MovieTopia
                             FROM
                                 Theatre t;";
 
-                adapter.SelectCommand = new SqlCommand(sql, conn); ;
+                adapter.SelectCommand = new SqlCommand(sql, conn);
                 adapter.Fill(ds, "Theatre");
 
-                dgvData.DataSource = ds;
-                dgvData.DataMember = "Theatre";
+                //dgvData.DataSource = ds;
+                //dgvData.DataMember = "Theatre";
+                dgvData.DataSource = ds.Tables[tblName].DefaultView;
             }
         }
 
         private void AdjustDataGridViewSize()
         {
             dgvData.Width = this.ClientSize.Width - (2 * padding);
-            dgvData.Height = this.ClientSize.Height - (10 * padding);
-            dgvData.Location = new Point(padding, padding * 3);
+            dgvData.Height = this.ClientSize.Height - (11 * padding);
+            dgvData.Location = new Point(padding, padding * 5);
         }
 
         private void AdjustColumnWidths()
@@ -93,7 +101,13 @@ namespace MovieTopia
 
         private void btnNew_Click(object sender, EventArgs e)
         {
-            DetailsForm detailsForm = new DetailsForm("Theatre", ds, null, null);
+            Dictionary<string, string> attributeNameMap = new Dictionary<string, string>();
+            attributeNameMap["TheatreID"] = "Theatre ID";
+            attributeNameMap["TheatreName"] = "Theatre Name";
+            attributeNameMap["NumRows"] = "Number of Rows";
+            attributeNameMap["NumCols"] = "Number of Columns";
+
+            DetailsForm detailsForm = new DetailsForm("Theatre", ds, null, null, attributeNameMap);
             DialogResult result = detailsForm.ShowDialog();
             if (result == DialogResult.OK)
             {
@@ -134,6 +148,10 @@ namespace MovieTopia
                         command.ExecuteNonQuery();
                         MessageBox.Show("Created Successfully", "Success");
                     }
+                    catch (SqlException)
+                    {
+                        MessageBox.Show("The entered Theatre name already exits", "Error");
+                    }
                     catch (Exception ex)
                     {
                         MessageBox.Show(ex.Message, "Error");
@@ -150,7 +168,13 @@ namespace MovieTopia
             {
                 DataGridViewRow selectedRow = dgvData.SelectedRows[0];
 
-                DetailsForm detailsForm = new DetailsForm("Theatre", ds, selectedRow, null);
+                Dictionary<string, string> attributeNameMap = new Dictionary<string, string>();
+                attributeNameMap["TheatreID"] = "Theatre ID";
+                attributeNameMap["TheatreName"] = "Theatre Name";
+                attributeNameMap["NumRows"] = "Number of Rows";
+                attributeNameMap["NumCols"] = "Number of Columns";
+
+                DetailsForm detailsForm = new DetailsForm("Theatre", ds, selectedRow, null, attributeNameMap);
                 DialogResult result = detailsForm.ShowDialog();
                 if (result == DialogResult.OK)
                 {
@@ -188,6 +212,10 @@ namespace MovieTopia
                             command.ExecuteNonQuery();
                             MessageBox.Show("Updated Successfully", "Success");
                         }
+                        catch (SqlException)
+                        {
+                            MessageBox.Show("The entered Theatre name already exits", "Error");
+                        }
                         catch (Exception ex)
                         {
                             MessageBox.Show(ex.Message, "Error");
@@ -197,6 +225,10 @@ namespace MovieTopia
                     LoadData();
                 }
             }
+            else
+            {
+                MessageBox.Show("Please select a Theatre to edit.");
+            }
         }
 
         private void btnDelete_Click(object sender, EventArgs e)
@@ -204,6 +236,9 @@ namespace MovieTopia
             if (dgvData.SelectedRows.Count == 1)
             {
                 DataGridViewRow selectedRow = dgvData.SelectedRows[0];
+
+                DialogResult confirm = MessageBox.Show($"Are you sure you want to delete \"{selectedRow.Cells["TheatreName"].Value}\"?", "Delete Theatre", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (confirm == DialogResult.No) return;
 
                 string sql = @"
                         DELETE FROM 
@@ -227,6 +262,10 @@ namespace MovieTopia
                         command.ExecuteNonQuery();
                         MessageBox.Show("Deleted Successfully", "Success");
                     }
+                    catch (SqlException)
+                    {
+                        MessageBox.Show($"\"{selectedRow.Cells["TheatreName"].Value}\" cannot be deleted because of 1 or more Scheduled Movies that are dependent on this Theatre. Please delete the corresponding records before attempting to delete this Theatre.", "Error");
+                    }
                     catch (Exception ex)
                     {
                         MessageBox.Show(ex.Message, "Error");
@@ -234,6 +273,54 @@ namespace MovieTopia
                 }
 
                 LoadData();
+            }
+            else
+            {
+                MessageBox.Show("Please select a Theatre to delete.");
+            }
+        }
+
+        private void btnReturn_Click(object sender, EventArgs e)
+        {
+            this.Close();
+        }
+
+        private void txtFilter_TextChanged(object sender, EventArgs e)
+        {
+            string filterText = txtFilter.Text;
+            DataTable dt = ds.Tables[tblName];
+
+            if (string.IsNullOrEmpty(filterText))
+            {
+                dt.DefaultView.RowFilter = "";
+            }
+            else
+            {
+                // Construct the filter string
+                var filterConditions = dt.Columns.Cast<DataColumn>()
+                    .Select(c => {
+                        if (c.DataType == typeof(string))
+                        {
+                            return $"{c.ColumnName} LIKE '%{filterText}%'";
+                        }
+                        else if (c.DataType == typeof(int) || c.DataType == typeof(decimal))
+                        {
+                            // Try parsing filterText to avoid applying invalid filter
+                            if (decimal.TryParse(filterText, out _))
+                            {
+                                return $"{c.ColumnName} = {filterText}";
+                            }
+                            else
+                            {
+                                return null;
+                            }
+                        }
+                        return null;
+                    })
+                    .Where(condition => condition != null); // Filter out any null conditions
+
+                // Combine all filter conditions using "OR"
+                dt.DefaultView.RowFilter = string.Join(" OR ", filterConditions);
             }
         }
     }
